+++ artifacts/kuhni-na-zakaz/.next/static/chunks/webpack.js
@@ -190,7 +190,7 @@
 /******/ 	
 /******/ 	/* webpack/runtime/getFullHash */
 /******/ 	(() => {
-/******/ 		__webpack_require__.h = () => ("620377124272f24f")
+/******/ 		__webpack_require__.h = () => ("fa98ec375e188dc8")
 /******/ 	})();
 /******/ 	
 /******/ 	/* webpack/runtime/global */